@@ -22,7 +22,7 @@
   
   <!-- Language + Compiler Settings-->
   <PropertyGroup>
-    <LangVersion>11</LangVersion>
+    <LangVersion>12</LangVersion>
   </PropertyGroup>
   
   <!--MISC-->
