@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-
-using Fido2NetLib;
+﻿using Fido2NetLib;
 
 namespace Fido2Demo;
 
