@@ -1,7 +1,4 @@
-﻿using System;
-using System.Net.Http;
-
-using Fido2NetLib;
+﻿using Fido2NetLib;
 
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection.Extensions;
