@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Fido2NetLib.Exceptions;
+﻿namespace Fido2NetLib.Exceptions;
 
 [Flags]
 public enum Fido2ErrorCode
