@@ -1,7 +1,4 @@
-﻿using System;
-using System.Runtime.Serialization;
-
-using Fido2NetLib.Exceptions;
+﻿using Fido2NetLib.Exceptions;
 
 namespace Fido2NetLib;
 
