@@ -52,16 +52,16 @@ internal sealed class FidoU2f : AttestationVerifier
         var y = (byte[])request.CredentialPublicKey[COSE.KeyTypeParameter.Y];
 
         // 4c.Let publicKeyU2F be the concatenation 0x04 || x || y
-        var publicKeyU2F = DataHelper.Concat(stackalloc byte[1] { 0x4 }, x, y);
+        byte[] publicKeyU2F = [0x4, .. x, .. y];
 
         // 5. Let verificationData be the concatenation of (0x00 || rpIdHash || clientDataHash || credentialId || publicKeyU2F)
-        byte[] verificationData = DataHelper.Concat(
-            stackalloc byte[1] { 0x00 },
-            request.AuthData.RpIdHash,
-            request.ClientDataHash,
-            request.AuthData.AttestedCredentialData.CredentialId,
-            publicKeyU2F
-        );
+        byte[] verificationData = [
+            0x00,
+            .. request.AuthData.RpIdHash,
+            .. request.ClientDataHash,
+            .. request.AuthData.AttestedCredentialData.CredentialId,
+            .. publicKeyU2F
+        ];
 
         // 6. Verify the sig using verificationData and certificate public key
         if (!request.TryGetSig(out byte[]? sig))
