@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Net.Http;
 using System.Security.Cryptography;
@@ -10,6 +9,7 @@ using System.Threading.Tasks;
 
 using Fido2NetLib.Serialization;
 
+using Microsoft.IdentityModel.JsonWebTokens;
 using Microsoft.IdentityModel.Tokens;
 
 namespace Fido2NetLib;
@@ -125,26 +125,26 @@ public sealed class Fido2MetadataServiceRepository : IMetadataRepository
         certChain.ChainPolicy.ExtraStore.Add(rootCert);
         certChain.ChainPolicy.RevocationMode = X509RevocationMode.NoCheck;
 
-        var validationParameters = new TokenValidationParameters
-        {
-            ValidateIssuer = false,
-            ValidateAudience = false,
-            ValidateLifetime = false,
-            ValidateIssuerSigningKey = true,
-            IssuerSigningKeys = blobPublicKeys
-        };
-
-        var tokenHandler = new JwtSecurityTokenHandler()
+        var tokenHandler = new JsonWebTokenHandler
         {
             // 250k isn't enough bytes for conformance test tool
             // https://github.com/AzureAD/azure-activedirectory-identitymodel-extensions-for-dotnet/issues/1097
             MaximumTokenSizeInBytes = rawBLOBJwt.Length
         };
 
-        tokenHandler.ValidateToken(
-            rawBLOBJwt,
-            validationParameters,
-            out var validatedToken);
+        var validateTokenResult = await tokenHandler.ValidateTokenAsync(rawBLOBJwt, new TokenValidationParameters
+        {
+            ValidateIssuer = false,
+            ValidateAudience = false,
+            ValidateLifetime = false,
+            ValidateIssuerSigningKey = true,
+            IssuerSigningKeys = blobPublicKeys
+        }).ConfigureAwait(false);
+
+        if (!validateTokenResult.IsValid)
+        {
+            throw new Fido2VerificationException("rawBLOBJwt is not valid");
+        }
 
         if (blobCerts.Length > 1)
         {
@@ -190,9 +190,9 @@ public sealed class Fido2MetadataServiceRepository : IMetadataRepository
         if (!certChainIsValid)
             throw new Fido2VerificationException("Failed to validate cert chain while parsing BLOB");
 
-        var blobPayload = ((JwtSecurityToken)validatedToken).Payload.SerializeToJson();
+        var blobPayload = ((JsonWebToken)validateTokenResult.SecurityToken).EncodedPayload;
 
-        MetadataBLOBPayload blob = JsonSerializer.Deserialize(blobPayload, FidoModelSerializerContext.Default.MetadataBLOBPayload)!;
+        MetadataBLOBPayload blob = JsonSerializer.Deserialize(Base64Url.Decode(blobPayload), FidoModelSerializerContext.Default.MetadataBLOBPayload)!;
         blob.JwtAlg = blobAlg;
         return blob;
     }
